implemented ARQ

+++ server/Packet.h
@@ -2,8 +2,8 @@
 #pragma once
 typedef struct Packet {
 	char txtCol[256];
-	unsigned short seqNr;
-	unsigned short checkSum;
+	long seqNr;
+	long checkSum;
 
 
 }packet;
